# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "LMS Backend Core API",
        "description": "Learning Management System Backend API Documentation. This API provides comprehensive endpoints for managing courses, users, enrollments, and learning progress.",
        "contact": {
            "name": "LMS API Support",
            "email": "api-support@lms.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "LMS API Server"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get all categories",
                "description": "Retrieve a list of all course categories",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "parameters": [
                    {
                        "name": "parent_id",
                        "in": "query",
                        "description": "Filter by parent category ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by category level",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Show only active categories",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categories retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Categories retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Programming"
                                                    },
                                                    "slug": {
                                                        "type": "string",
                                                        "example": "programming"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Programming and software development courses"
                                                    },
                                                    "icon": {
                                                        "type": "string",
                                                        "example": "fas fa-code"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://example.com/images/programming.jpg"
                                                    },
                                                    "parent_id": {
                                                        "type": "integer",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "level": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "sort_order": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "is_active": {
                                                        "type": "boolean",
                                                        "example": true
                                                    },
                                                    "courses_count": {
                                                        "type": "integer",
                                                        "example": 25
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "children": {
                                                        "type": "array",
                                                        "items": {
                                                            "$ref": "#/components/schemas/Category"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category details",
                "description": "Retrieve details of a specific category",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Category"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Category not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/categories/{id}/courses": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get courses in category",
                "description": "Retrieve courses belonging to a specific category",
                "operationId": "dda17c38523e1b427bb23d19cf37ed63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 12
                        }
                    },
                    {
                        "name": "level",
                        "in": "query",
                        "description": "Filter by course level",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "beginner",
                                "intermediate",
                                "advanced"
                            ]
                        }
                    },
                    {
                        "name": "price_type",
                        "in": "query",
                        "description": "Filter by price type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "free",
                                "paid"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "newest",
                            "enum": [
                                "newest",
                                "oldest",
                                "popular",
                                "title",
                                "price"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Courses retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Courses retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 12
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Course"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "title": "Course",
                "description": "Course model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Introduction to Laravel"
                    },
                    "slug": {
                        "type": "string",
                        "example": "introduction-to-laravel"
                    },
                    "short_description": {
                        "type": "string",
                        "example": "Learn Laravel framework basics"
                    },
                    "description": {
                        "type": "string",
                        "example": "Comprehensive course covering Laravel fundamentals..."
                    },
                    "thumbnail_url": {
                        "type": "string",
                        "example": "https://example.com/thumbnails/course1.jpg"
                    },
                    "level": {
                        "type": "string",
                        "enum": [
                            "beginner",
                            "intermediate",
                            "advanced"
                        ],
                        "example": "beginner"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "published"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 99.99
                    },
                    "estimated_duration_hours": {
                        "type": "integer",
                        "example": 20
                    },
                    "language": {
                        "type": "string",
                        "example": "en"
                    },
                    "is_featured": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_free": {
                        "type": "boolean",
                        "example": false
                    },
                    "published_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "instructor": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "enrollment_count": {
                        "type": "integer",
                        "example": 150
                    },
                    "total_chapters": {
                        "type": "integer",
                        "example": 12
                    }
                },
                "type": "object"
            },
            "Category": {
                "title": "Category",
                "description": "Course category model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Programming"
                    },
                    "slug": {
                        "type": "string",
                        "example": "programming"
                    },
                    "description": {
                        "type": "string",
                        "example": "Programming and software development courses"
                    },
                    "icon": {
                        "type": "string",
                        "example": "fas fa-code"
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/images/programming.jpg"
                    },
                    "parent_id": {
                        "type": "integer",
                        "example": null,
                        "nullable": true
                    },
                    "level": {
                        "type": "integer",
                        "example": 1
                    },
                    "sort_order": {
                        "type": "integer",
                        "example": 1
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "courses_count": {
                        "type": "integer",
                        "example": 25
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in format: Bearer <token>"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}